+++ src/chat.back/Chat/Chat.API/Controllers/FileController.cs
@@ -8,29 +8,12 @@ namespace Chat.API.Controllers;
 
 [Route("api/files")]
 [ApiController]
-public class FileController : Controller
+public class FileController : ControllerBase
 {
-    //private readonly S3Client _s3Client;
     private readonly IAmazonS3 _s3Client;
     public FileController(IAmazonS3 s3Client)
     {
         _s3Client = s3Client;
-    }
-    
-    [HttpPost("create-bucket")]
-    public async Task<IActionResult> CreateBucketAsync(string bucketName)
-    {
-        var bucketExists = await _s3Client.DoesS3BucketExistAsync(bucketName);
-        if (bucketExists) return BadRequest($"Bucket {bucketName} already exists.");
-        await _s3Client.PutBucketAsync(bucketName);
-        return Ok($"Bucket {bucketName} created.");
-    }
-    
-    [HttpDelete("delete-bucket")]
-    public async Task<IActionResult> DeleteBucketAsync(string bucketName)
-    {
-        await _s3Client.DeleteBucketAsync(bucketName);
-        return NoContent();
     }
     
     [HttpPost("upload")]
